requirements.txt
# Core Python Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
pyyaml>=6.0
aiohttp>=3.8.0

# AI & ML
openai==1.3.5
anthropic==0.7.7
google-generativeai>=0.3.0
langchain==0.0.340
langchain-community==0.0.1
langchain-core==0.1.0
transformers==4.35.0
torch>=2.0.0
sentence-transformers==2.2.2
chromadb==0.4.15
ollama==0.1.7

# Voice & Audio
openai-whisper==20231117
TTS==0.20.6
pyaudio==0.2.11
sounddevice==0.4.6
librosa==0.10.1

# Vision & Image Processing
opencv-python==4.8.1.78
pillow==10.1.0
clip-by-openai==1.0

# Web Scraping & HTTP
requests==2.31.0
httpx==0.25.2
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# Task Queue & Background Jobs
celery==5.3.4
redis==5.0.1
dramatiq==1.15.0

# Database & Storage
psycopg2-binary==2.9.9
pymongo==4.6.0
elasticsearch==8.11.0

# Security & Encryption
cryptography==41.0.7
bcrypt==4.1.1
pyotp==2.9.0
pynacl==1.5.0

# System Integration
pyautogui==0.9.54
psutil==5.9.6
watchdog==3.0.0
schedule==1.2.0
python-crontab==3.0.0

# Notification & Communication
plyer==2.1.0
smtplib-ssl==1.0.0
twilio==8.10.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pyyaml==6.0.1
toml==0.10.2
jinja2==3.1.2
aiofiles==23.2.1

# Platform-specific (Windows)
pywin32==306; sys_platform == "win32"
wmi==1.5.1; sys_platform == "win32"

# Monitoring & Logging
structlog==23.2.0
sentry-sdk==1.38.0
prometheus-client==0.19.0
